+++ crates/objc2/CHANGELOG.md
@@ -117,6 +117,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * **BREAKING**: Mark `Retained::autorelease` as `unsafe`, since we cannot
   ensure that the given pool is actually the innermost pool.
 * **BREAKING**: Removed the deprecated `malloc` feature and `malloc_buf` dependency.
+* **BREAKING**: Removed aliases `DefaultId`, `IdFromIterator` and
+  `IdIntoIterator`, as well as their methods. Use the renamed traits instead.
 
 ### Fixed
 * Remove an incorrect assertion when adding protocols to classes in an unexpected